+++ tests/Unit/Classes/ConstantsHelperTest.php
@@ -22,6 +22,24 @@ class ConstantsHelperTest extends TestCase
             "FIRST_CONSTANT" => 1,
             "SECOND_CONSTANT" => 2,
             "THIRD_CONSTANT" => 3,
-        ], $class->getConstants());
+        ], $class::getConstants());
+    }
+
+    public function test it returns all constants through the shorthand method()
+    {
+        $class = new class
+        {
+            use Constants;
+
+            const FIRST_CONSTANT = 1;
+            const SECOND_CONSTANT = 2;
+            const THIRD_CONSTANT = 3;
+        };
+
+        $this->assertEquals([
+            "FIRST_CONSTANT" => 1,
+            "SECOND_CONSTANT" => 2,
+            "THIRD_CONSTANT" => 3,
+        ], $class::constants());
     }
 }