+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Set up testing environment and add tests for each feature
 - Added support for Laravel 5.5
 - Added `sss_if` global helper
+- Added a shorthand `constants()` method to the `Constants` trait
 
 ### Changed